fixup! JBAI-13946 Extend integration tests to cover new models and move to separate module
