+++ integration-tests/build.gradle.kts
@@ -20,7 +20,6 @@ kotlin {
                 implementation(kotlin("test"))
                 implementation(kotlin("test-junit5"))
                 implementation("org.junit.jupiter:junit-jupiter-params:5.9.2")
-                implementation(libs.ai.grazie.utils.common)
                 implementation(libs.kotlinx.coroutines.test)
                 implementation(libs.ktor.client.content.negotiation)
             }
