+++ integration-tests/src/jvmTest/kotlin/ai/jetbrains/code/integration/tests/KotlinAIAgentWithMultipleLLMIntegrationTest.kt
@@ -1,31 +1,31 @@
 package ai.jetbrains.code.integration.tests
 
-import ai.grazie.code.agents.core.agent.AIAgent
-import ai.grazie.code.agents.core.agent.AIAgentException
-import ai.grazie.code.agents.core.agent.config.AIAgentConfig
-import ai.grazie.code.agents.core.dsl.builder.forwardTo
-import ai.grazie.code.agents.core.dsl.builder.strategy
-import ai.grazie.code.agents.core.dsl.extension.*
-import ai.grazie.code.agents.core.tools.*
-import ai.grazie.code.agents.ext.agent.simpleSingleRunAgent
-import ai.grazie.code.agents.local.features.eventHandler.feature.EventHandler
-import ai.grazie.code.agents.local.features.eventHandler.feature.EventHandlerConfig
-import ai.grazie.code.agents.local.features.tracing.feature.Tracing
 import ai.jetbrains.code.integration.tests.ReportingLLMLLMClient.Event
 import ai.jetbrains.code.integration.tests.TestUtils.readTestAnthropicKeyFromEnv
 import ai.jetbrains.code.integration.tests.TestUtils.readTestOpenAIKeyFromEnv
-import ai.jetbrains.code.prompt.dsl.Prompt
-import ai.jetbrains.code.prompt.dsl.prompt
-import ai.jetbrains.code.prompt.executor.clients.LLMClient
-import ai.jetbrains.code.prompt.executor.clients.anthropic.AnthropicLLMClient
-import ai.jetbrains.code.prompt.executor.clients.anthropic.AnthropicModels
-import ai.jetbrains.code.prompt.executor.clients.openai.OpenAILLMClient
-import ai.jetbrains.code.prompt.executor.clients.openai.OpenAIModels
-import ai.jetbrains.code.prompt.executor.llms.MultiLLMPromptExecutor
-import ai.jetbrains.code.prompt.executor.llms.all.simpleAnthropicExecutor
-import ai.jetbrains.code.prompt.llm.LLMProvider
-import ai.jetbrains.code.prompt.llm.LLModel
-import ai.jetbrains.code.prompt.message.Message
+import ai.koog.agents.core.agent.AIAgent
+import ai.koog.agents.core.agent.AIAgentException
+import ai.koog.agents.core.agent.config.AIAgentConfig
+import ai.koog.agents.core.dsl.builder.forwardTo
+import ai.koog.agents.core.dsl.builder.strategy
+import ai.koog.agents.core.dsl.extension.*
+import ai.koog.agents.core.tools.*
+import ai.koog.agents.ext.agent.simpleSingleRunAgent
+import ai.koog.agents.local.features.eventHandler.feature.EventHandler
+import ai.koog.agents.local.features.eventHandler.feature.EventHandlerConfig
+import ai.koog.agents.local.features.tracing.feature.Tracing
+import ai.koog.prompt.dsl.Prompt
+import ai.koog.prompt.dsl.prompt
+import ai.koog.prompt.executor.clients.LLMClient
+import ai.koog.prompt.executor.clients.anthropic.AnthropicLLMClient
+import ai.koog.prompt.executor.clients.anthropic.AnthropicModels
+import ai.koog.prompt.executor.clients.openai.OpenAILLMClient
+import ai.koog.prompt.executor.clients.openai.OpenAIModels
+import ai.koog.prompt.executor.llms.MultiLLMPromptExecutor
+import ai.koog.prompt.executor.llms.all.simpleAnthropicExecutor
+import ai.koog.prompt.llm.LLMProvider
+import ai.koog.prompt.llm.LLModel
+import ai.koog.prompt.message.Message
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.channels.Channel
