+++ integration-tests/src/jvmTest/kotlin/ai/jetbrains/code/integration/tests/MultipleLLMPromptExecutorIntegrationTest.kt
@@ -1,21 +1,21 @@
 package ai.jetbrains.code.integration.tests
 
-import ai.grazie.code.agents.core.tools.ToolDescriptor
-import ai.grazie.code.agents.core.tools.ToolParameterDescriptor
-import ai.grazie.code.agents.core.tools.ToolParameterType
 import ai.jetbrains.code.integration.tests.Models.modelsWithoutToolsSupport
 import ai.jetbrains.code.integration.tests.TestUtils.readTestAnthropicKeyFromEnv
 import ai.jetbrains.code.integration.tests.TestUtils.readTestGoogleAIKeyFromEnv
 import ai.jetbrains.code.integration.tests.TestUtils.readTestOpenAIKeyFromEnv
-import ai.jetbrains.code.prompt.dsl.Prompt
-import ai.jetbrains.code.prompt.executor.clients.anthropic.AnthropicLLMClient
-import ai.jetbrains.code.prompt.executor.clients.google.GoogleLLMClient
-import ai.jetbrains.code.prompt.executor.clients.openai.OpenAILLMClient
-import ai.jetbrains.code.prompt.executor.llms.MultiLLMPromptExecutor
-import ai.jetbrains.code.prompt.executor.llms.all.DefaultMultiLLMPromptExecutor
-import ai.jetbrains.code.prompt.llm.LLMProvider
-import ai.jetbrains.code.prompt.llm.LLModel
-import ai.jetbrains.code.prompt.message.Message
+import ai.koog.agents.core.tools.ToolDescriptor
+import ai.koog.agents.core.tools.ToolParameterDescriptor
+import ai.koog.agents.core.tools.ToolParameterType
+import ai.koog.prompt.dsl.Prompt
+import ai.koog.prompt.executor.clients.anthropic.AnthropicLLMClient
+import ai.koog.prompt.executor.clients.google.GoogleLLMClient
+import ai.koog.prompt.executor.clients.openai.OpenAILLMClient
+import ai.koog.prompt.executor.llms.MultiLLMPromptExecutor
+import ai.koog.prompt.executor.llms.all.DefaultMultiLLMPromptExecutor
+import ai.koog.prompt.llm.LLMProvider
+import ai.koog.prompt.llm.LLModel
+import ai.koog.prompt.message.Message
 import kotlinx.coroutines.flow.toList
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.Assumptions.assumeTrue
