+++ integration-tests/src/jvmTest/kotlin/ai/jetbrains/code/integration/tests/OpenAIEmbeddingsIntegrationTest.kt
@@ -1,8 +1,8 @@
 package ai.jetbrains.code.integration.tests
 
 import ai.jetbrains.code.integration.tests.TestUtils.readTestOpenAIKeyFromEnv
-import ai.jetbrains.code.prompt.executor.clients.openai.OpenAILLMClient
-import ai.jetbrains.code.prompt.executor.clients.openai.OpenAIModels
+import ai.koog.prompt.executor.clients.openai.OpenAILLMClient
+import ai.koog.prompt.executor.clients.openai.OpenAIModels
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.Test
 import kotlin.test.assertNotNull
