+++ integration-tests/src/jvmTest/kotlin/ai/jetbrains/code/integration/tests/SimpleAgentIntegrationTest.kt
@@ -1,20 +1,20 @@
 package ai.jetbrains.code.integration.tests
 
-import ai.grazie.code.agents.core.tools.ToolRegistry
-import ai.grazie.code.agents.ext.agent.simpleChatAgent
-import ai.grazie.code.agents.ext.agent.simpleSingleRunAgent
-import ai.grazie.code.agents.ext.tool.SayToUser
-import ai.grazie.code.agents.local.features.eventHandler.feature.EventHandler
-import ai.grazie.code.agents.local.features.eventHandler.feature.EventHandlerConfig
 import ai.jetbrains.code.integration.tests.Models.modelsWithoutToolsSupport
 import ai.jetbrains.code.integration.tests.TestUtils.readTestAnthropicKeyFromEnv
 import ai.jetbrains.code.integration.tests.TestUtils.readTestGoogleAIKeyFromEnv
 import ai.jetbrains.code.integration.tests.TestUtils.readTestOpenAIKeyFromEnv
-import ai.jetbrains.code.prompt.executor.llms.all.simpleAnthropicExecutor
-import ai.jetbrains.code.prompt.executor.llms.all.simpleGoogleAIExecutor
-import ai.jetbrains.code.prompt.executor.llms.all.simpleOpenAIExecutor
-import ai.jetbrains.code.prompt.llm.LLMProvider
-import ai.jetbrains.code.prompt.llm.LLModel
+import ai.koog.agents.core.tools.ToolRegistry
+import ai.koog.agents.ext.agent.simpleChatAgent
+import ai.koog.agents.ext.agent.simpleSingleRunAgent
+import ai.koog.agents.ext.tool.SayToUser
+import ai.koog.agents.local.features.eventHandler.feature.EventHandler
+import ai.koog.agents.local.features.eventHandler.feature.EventHandlerConfig
+import ai.koog.prompt.executor.llms.all.simpleAnthropicExecutor
+import ai.koog.prompt.executor.llms.all.simpleGoogleAIExecutor
+import ai.koog.prompt.executor.llms.all.simpleOpenAIExecutor
+import ai.koog.prompt.llm.LLMProvider
+import ai.koog.prompt.llm.LLModel
 import kotlinx.coroutines.runBlocking
 import org.junit.jupiter.api.Assumptions.assumeTrue
 import org.junit.jupiter.api.Disabled
